+++ framework/src/main/java/edu/uci/eecs/wukong/framework/util/PipelineUtil.java
@@ -7,7 +7,7 @@ import edu.uci.eecs.wukong.framework.annotation.WuTimer;
 public class PipelineUtil {
 	public final static int DEFAULT_INTERVAL = 5;
 	
-	public static int getIntervalFromAnnotation(Object object) {
+	public static float getIntervalFromAnnotation(Object object) {
 		Annotation[] annotations = object.getClass().getAnnotations();
 		for (Annotation annotation : annotations) {
 			if (annotation.annotationType().equals(WuTimer.class)) {
